fix: standardize dtos

+++ src/auth/auth.service.ts
@@ -1,6 +1,7 @@
 import {
   BadRequestException,
   ConflictException,
+  ForbiddenException,
   Injectable,
 } from '@nestjs/common';
 import { InjectConnection, InjectModel } from '@nestjs/mongoose';
@@ -10,6 +11,7 @@ import { CreateAuthDto } from './dto/create-auth.dto';
 import { UserService } from '../user/user.service';
 import { RegistrationAuthDto } from './dto/registration-auth.dto';
 import { AuthDocument, AuthModel, AuthType } from './entities/auth.entity';
+import { UserRole } from '../user/entities/user.entity';
 
 @Injectable()
 export class AuthService {
@@ -31,15 +33,30 @@ export class AuthService {
 
   async signUpUser(registrationDto: RegistrationAuthDto) {
     if (registrationDto.type === AuthType.Password && !registrationDto.email) {
-      throw new BadRequestException(`${AuthType.Password}::EMAIL_NOT_PROVIDED`);
+      throw new BadRequestException(`AUTH::SIGNUP::EMAIL_NOT_PROVIDED`);
     }
 
     let user;
 
     const session = await this.connection.startSession();
     await session.withTransaction(async () => {
-      user = await this.userService.create(registrationDto);
-      await this.createAuthEntity(registrationDto);
+      // create user first
+      user = await this.userService.create({
+        username: registrationDto.username,
+        email: registrationDto.email,
+        displayName: registrationDto.displayName,
+        avatar: registrationDto.avatar,
+        isEnabled: true,
+        isEmailVerified: false,
+        role: [UserRole.User],
+      });
+
+      // then create auth entity
+      await this.createAuthEntity({
+        type: registrationDto.type,
+        credentials: registrationDto.credentials,
+        userId: user.id,
+      });
     });
 
     session.endSession();
@@ -51,7 +68,7 @@ export class AuthService {
       createAuthDto.type === AuthType.Password &&
       (await this.AuthDocument.findOne({ type: createAuthDto.type }))
     ) {
-      throw new ConflictException(`${AuthType.Password}::DUPLICATED_ENTITIES`);
+      throw new ConflictException(`AUTH::CREATE::DUPLICATED_ENTITIES`);
     }
 
     const authDocument = new this.AuthDocument({
@@ -63,7 +80,15 @@ export class AuthService {
     return authDocument.save();
   }
 
-  deleteAuthEntity(id: string) {
+  async deleteAuthEntity(id: string) {
+    const entity = await this.AuthDocument.findById(id);
+
+    if ((await this.AuthDocument.count({ userId: entity.userId })) === 1) {
+      throw new ForbiddenException(
+        'AUTH::DELETE::AT_LEAST_ONE_ENTITY_CONSTRAINT',
+      );
+    }
+
     return this.AuthDocument.findByIdAndRemove(id);
   }
 }
